Create The beginning of the rise
After losing everything I had built in the last 18 years. 18 years after finding my father After he committed suicide.... The destroying angel was born,"Todesengel" also known as the angel of death. Or the Grimm Reaper as many have came to know him by. I want to say that the Angel of death is not a angry angel or a violent angel. Rather he is a angel that brings peace and rest for those who meet him. He greets them on the other side, and his job is to make sure his new friends are comfortable and get what needed rest and peace they have been seeking.